refactor: 메인페이지 타입 선언 분리
