+++ src/api/main.ts
@@ -1,21 +1,7 @@
 import { client } from '@/api/client';
+import type { Award, MainEvent } from '@/types/Main';
 
-// 타입 정의
-
-export interface MainEvent {
-  id: number;
-  title: string;
-  start: string;
-  end: string;
-  imgUrl: string | null;
-}
-
-export interface Award {
-  tabIdx: number;
-  year: number;
-  type: [number, number, number, number];
-}
-
+// API 함수
 export const getMainEvents = () => {
   return client.get<{ data: MainEvent[] }>('/api/v1/events').then((res) => res.data.data);
 };
